+++ tests/test_versioning_unittest.py
@@ -1,6 +1,5 @@
 import subprocess
 import unittest
-from importlib.metadata import PackageNotFoundError
 from pathlib import Path
 from unittest.mock import patch
 
@@ -11,7 +10,6 @@ from drain_swamp.version_semantic import (
     _current_tag,
     _current_version,
     _get_app_name,
-    _is_setuptools_scm,
     _map_release,
     _remove_v,
     get_version,
@@ -57,18 +55,6 @@ testdata_releaselevel = (
     ("b", "beta"),
     ("beta", "beta"),
 )
-
-
-class PackageInstalled(unittest.TestCase):
-    def test_is_setuptools_scm(self):
-        """Checks whether or not package setuptools-scm is installed"""
-        # non-existent nor installed package
-        with patch(
-            "importlib.metadata.version",
-            side_effect=PackageNotFoundError,
-        ):
-            bol_actual = _is_setuptools_scm()
-            self.assertFalse(bol_actual)
 
 
 class PackageVersioning(unittest.TestCase):
@@ -331,7 +317,7 @@ class SemVersioning(unittest.TestCase):
         # setuptools-scm is not installed.
         #    Current version is obtained from package setuptools-scm
         with patch(
-            f"{g_app_name}.version_semantic._is_setuptools_scm",
+            f"{g_app_name}.version_semantic.is_package_installed",
             return_value=False,
         ):
             actual = _current_version(path=self.cwd)